requirements.txt
mtcnn
tensorflow
keras
keras-vggface
keras_applications
PyYAML
tqdm
scikit-learn
streamlit
bing-image-downloader

#local packages-
-e .